refactor: extract logic to use-game custom hook
